# Phase 2-20260109T205517Z-3-001/Phase 2/Phase 2/swagger-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ComStag API",
    "description": "Organizations & Consumers platform",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://comstag-back.onrender.com",
      "description": "Generated server url"
    }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/v1/profile/reset-pass": {
      "put": {
        "tags": ["profile-reset-password-processor"],
        "summary": "Reset profile password",
        "description": "Reset password for the current user\n",
        "operationId": "resetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/v1/event/{eventId}": {
      "get": {
        "tags": ["get-event-processor"],
        "summary": "Get event by id",
        "operationId": "getEvent",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/EventDm" } }
            }
          }
        }
      },
      "put": {
        "tags": ["update-event-processor"],
        "summary": "Update event",
        "description": "1. Update specific event by id.\n2. Only organization have access to this endpoint.\n3. Upload media first if needed\n",
        "operationId": "editEvent",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateEventRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      },
      "delete": {
        "tags": ["delete-event-processor"],
        "summary": "Delete event",
        "description": "1. This endpoint allow active organizations to delete event.\n",
        "operationId": "deleteEvent",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/v1/certificate/{certificateId}": {
      "put": {
        "tags": ["update-certificate-processor"],
        "summary": "Update certificate",
        "description": "Use this endpoint to allow organizations to create there certificates.\nIf the certificate have image id, the image will be uploaded to the storage first by upload certificate endpoint.\n",
        "operationId": "editCertificate",
        "parameters": [
          {
            "name": "certificateId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCertificateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      },
      "delete": {
        "tags": ["delete-certificate-processor"],
        "summary": "Delete certificate",
        "description": "Use this endpoint to allow organizations to delete there certificates",
        "operationId": "deleteCertificate",
        "parameters": [
          {
            "name": "certificateId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
...
    "/v1/event": {
      "post": {
        "tags": ["create-event-processor"],
        "summary": "Create event",
        "description": "1. This endpoint allow active organizations to create event.\n",
        "operationId": "createEvent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EventRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
...
    "/v1/certificate": {
      "post": {
        "tags": ["create-certificate-processor"],
        "summary": "Create certificate",
        "description": "Use this endpoint to allow organizations to create there certificates.\nIf the certificate have image id, the image will be uploaded to the storage first by upload certificate endpoint.\n",
        "operationId": "createCertificate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CertificateRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
...
    "/v1/auth/reset-pass": {
      "post": {
        "tags": ["auth-reset-password-processor"],
        "operationId": "resetPassword_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/v1/auth/refresh-token": {
      "post": {
        "tags": ["auth-refresh-token-processor"],
        "summary": "Refresh token",
        "description": "Use this endpoint to refresh token by sending refresh token",
        "operationId": "refreshToken",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": ["auth-logout-processor"],
        "summary": "Logout user",
        "description": "Use this endpoint to logout user by sending refresh token",
        "operationId": "logout",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": ["auth-login-processor"],
        "summary": "Login with username/password and start verification flow",
        "description": "1. Send username and password.\n2. If credentials are valid:\n   - The API returns the userId in the response.\n   - A 6-digit verification code is sent to the user's email.\n3. The returned userId must be used later in the Verify Code endpoint\n   together with the 6-digit code received via email.\n4. There is no need to authenticate this endpoint.\n",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthLoginRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": { "schema": { "type": "string", "format": "uuid" } }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProfileResetPasswordRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "newPassword": { "type": "string", "minLength": 1 },
          "oldPassword": { "type": "string", "minLength": 1 }
        },
        "required": ["newPassword", "oldPassword"]
      },
...
      "UpdateEventRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "body": { "type": "string" },
          "industryId": { "type": "integer", "format": "int64" },
          "country": { "type": "string" },
          "state": { "type": "string" },
          "city": { "type": "string" },
          "address": { "type": "string" },
          "startAt": { "type": "string", "format": "date-time" },
          "endAt": { "type": "string", "format": "date-time" },
          "online": { "type": "boolean" },
          "onlineLink": { "type": "string" },
          "deletedMediaIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "uniqueItems": true
          },
          "newMediaIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "uniqueItems": true
          }
        }
      },
      "UpdateCertificateRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "body": { "type": "string" },
          "link": { "type": "string" },
          "certificateDate": { "type": "string", "format": "date" },
          "newImageId": { "type": "string", "format": "uuid" }
        }
      },
...
      "EventRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 1 },
          "body": { "type": "string" },
          "industry": { "type": "integer", "format": "int64" },
          "country": { "type": "string" },
          "state": { "type": "string" },
          "city": { "type": "string" },
          "address": { "type": "string" },
          "startAt": { "type": "string", "format": "date-time" },
          "endAt": { "type": "string", "format": "date-time" },
          "online": { "type": "boolean" },
          "onlineLink": { "type": "string" },
          "mediaIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "uniqueItems": true
          }
        },
        "required": ["startAt", "title"]
      },
      "CertificateRequest": {
        "type": "object",
        "properties": {
          "imageId": { "type": "string", "format": "uuid" },
          "title": { "type": "string", "minLength": 1 },
          "body": { "type": "string" },
          "link": { "type": "string" },
          "certificateDate": { "type": "string", "format": "date" }
        },
        "required": ["certificateDate", "imageId", "title"]
      },
...
      "AuthResetPasswordRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "newPassword": { "type": "string", "minLength": 1 },
          "verificationCode": { "type": "string", "minLength": 1 }
        },
        "required": ["newPassword", "verificationCode"]
      },
...
      "AuthLoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 1 }
        },
        "required": ["password"]
      },
...
      "MediaDm": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "mediaType": { "type": "string", "enum": ["IMAGE", "VIDEO", "FILE"] }
        }
      },
...
      "IndustryDm": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" }
        }
      },
...
      "EventDm": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "orgId": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "body": { "type": "string" },
          "mode": { "type": "string", "enum": ["ONLINE", "OFFLINE"] },
          "industry": { "$ref": "#/components/schemas/IndustryDm" },
          "country": { "type": "string" },
          "state": { "type": "string" },
          "city": { "type": "string" },
          "address": { "type": "string" },
          "startAt": { "type": "string", "format": "date-time" },
          "endAt": { "type": "string", "format": "date-time" },
          "onlineLink": { "type": "string" },
          "viewers": { "type": "integer", "format": "int64" },
          "registeredCount": { "type": "integer", "format": "int64" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "media": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MediaDm" }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Paste your IdToken here",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}